Temporarily rename the change-event-result organization relationship (#94)
This relationship uses the same `org:organization` predicate that is also used for the local involvements relationship. When querying for local involvements data when an organization also has change events data,  mu-cl-resources receives the change events data instead and runs into errors. This is likely a bug in the micro service itself. Renaming the predicate is a quick workaround so we have more time to find the real solution for the problem without a broken app.