fix: removed write access to shared graph
This access was required to edit the (legal) name of municipalities and
provinces. This did lead to an issue were data for administrative units was
incorrectly written to both graphs. To avoid this we decided that, at least for
the moment, the name of municipalities and provinces is not editable, removing
the need to be able to write to the shared graph.