refactor: renamed units index to organizations
Soon we want to introduce (additional) non-administrative unit
organizations. This renaming makes it more clear this index is not limited to
administrative units, both here as where it is used in the frontend. Note,
technically the 'units' name was already incorrect as a 'Representatief orgaan'
is not an administrative unit.