fix: remove faulty address data for some AGBs
The migrations importing the public OCMW association's address data contained a
faulty query that matched on empty KBO numbers. As a result address data was
incorrectly inserted to some AGBs that also do not have a KBO number.

This migration removes the incorrect data, which as inserted into the
triplestore next to actual address. Furthermore, this also (re-)inserts the
correct provinces for the affected AGBs.