feat: mandates and functions for governing bodies legislature 2025
For each of the governing bodies created by the migrations in the previous
commit, now add the corresponding mandates and functions.

Within a organisation some governing bodies share the same mandates. For
example, for municipalities the governing body "Burgemeester" and "College van
burgemeester en schepen" both refer to the SAME instance of mandates for
"Burgemeester" and "Aangewezen burgemeester".
To achieve this the some migrations contain a second `INSERT` clause in which
the other governing body that contains the necessary mandates is selected to be
able to determine the correct mandate URI. This only works on two conditions:
1. Migrations that depend on each must be executed in the correct order. For
   example, the migration inserting the mandates for "College van burgemeester
   en schepen" must be executed after that for "Burgemeester". The former will
   reuse mandates created by the latter.
2. The UUID for the shared mandates is calculated in exactly the same way in
   migrations that depend on each other.