chore: enable editing contact data
Re-enable editing contact data in OP as the CLB app editing functionality was
disable due to data flow issues. This sets the relevant frontend feature flag
such that it allows editing the contact data for all organizations
classifications agains.

When deploying make sure this environment variable is **not** overwritten in the
environment's `docker-compose.override.yml`.