feat: include URI in responses for organisations
This includes an organisation's URI in the response from the resource
service. This allows to easily get the URI for an organisation without having to
determine it based to UUID and classification and/or explicitly querying the
triplestore.

This is mostly a convenience for developers, but can be leveraged to add a link
to Centrale Vindplaats in the frontend.